Rename taxonomy filters into filter pairs
